feat: Graphql context(current_user, trace_id) and error adapter
