+++ docs/content/docs/core-concepts/graphql.md
@@ -47,6 +47,77 @@ in the HTTP status.
 The 400 status is reserved for *transport-level* failures: malformed JSON in
 the POST body or an invalid query string in GET. Unsupported HTTP methods return 405.
 
- ### Limitations
+## Request context
 
-This pr just delivers the extractors, Schema management and the `with_graphql` builder method are still to be done!
+`RapinaGraphQLContext` carries Rapina's request-scoped data into resolvers
+the authenticated user and the trace id:
+
+```rust
+pub struct RapinaGraphQLContext {
+    pub current_user: Option<CurrentUser>,
+    pub trace_id: String,
+}
+```
+
+async-graphql's `Context` is a type-map, so resolvers retrieve it with
+`ctx.data::<RapinaGraphQLContext>()`:
+
+```rust
+use rapina::prelude;
+
+async fn me(&self, ctx: &Context<'_>) -> async_graphql::Result<User> {
+    let rapina_ctx = ctx.data::<RapinaGraphQLContext>()?;
+    let user = rapina_ctx.current_user.as_ref().ok_or_else(|| {
+        graphql_error("UNAUTHORIZED", "Login required", &rapina_ctx.trace_id)
+    })?;
+    // ...
+}
+```
+
+Automatic per-request injection  pulling `CurrentUser` and `trace_id` from the
+request and calling `request.data(ctx)` for you  arrives with the `with_graphql`
+builder method. Until then, inject it yourself before executing:
+
+```rust
+GraphQLResponse(schema.execute(req.0.data(rapina_ctx)).await)
+```
+
+## Errors
+
+Resolver errors can carry Rapina's error vocabulary  the same `code` and
+`trace_id`, the REST error envelope guarantees  into the GraphQL response's
+`errors[].extensions`. There are two entry points.
+
+**Bridging an existing `rapina::Error`.** It implements async-graphql's
+`ErrorExtensions`, so `.extend()` turns one into a GraphQL error that keeps its
+`code` (and `trace_id`, when the error carries one):
+
+```rust
+let post = find_post(id).await.extend()?;
+```
+
+> A bare `?` on a `Result<_, rapina::Error>` also compiles  async-graphql ships a
+> blanket `From<T: Display>`  but that path carries only the *message* and drops
+> `code`/`trace_id`. Use `.extend()` (or `map_err(|e| e.extend())`) when you want
+> the extensions.
+
+**Constructing an error inline.** When the error originates in the resolver and
+you have no `rapina::Error` to bridge, `graphql_error` builds one with both `code`
+and the current `trace_id` set:
+
+```rust
+let post = find_post(id).await.ok_or_else(|| {
+    graphql_error("NOT_FOUND", "Post not found", &rapina_ctx.trace_id)
+})?;
+```
+
+## Status
+
+Available today: the `GraphQLRequest`/`GraphQLResponse` extractor and responder,
+the `RapinaGraphQLContext` type, and the error bridge.
+
+Still to come:
+
+- **Automatic context injection** and **schema management** via a `with_graphql`
+  builder method  for now you wire the schema and inject the context manually.
+- **GraphiQL** playground.
